inicio do dashboard

+++ routes/api.php
@@ -21,7 +21,7 @@ Route::get('/items', 'ItemControllerAPI@index');
 
 //Route::get('departments', 'DepartmentControllerAPI@index');
 Route::get('/shift', 'ShiftControllerAPI@index');
-Route::get('users', 'UserControllerAPI@index');
+Route::middleware('auth.manager')->get('users', 'UserControllerAPI@index');
 Route::get('users/emailavailable', 'UserControllerAPI@emailAvailable');
 Route::middleware('auth:api')->get('users/me', 'UserControllerAPI@myProfile');
 
@@ -31,7 +31,7 @@ Route::get('users/{id}', 'UserControllerAPI@show');
 //Route::middleware('auth:api')->post('users', 'UserControllerAPI@store');
 
 Route::put('users/{id}', 'UserControllerAPI@update');
-Route::delete('users/{id}', 'UserControllerAPI@destroy');
+Route::middleware('auth.manager')->delete('users/{id}', 'UserControllerAPI@destroy');
 
 Route::post('upload/{id}', 'UserControllerAPI@upload')->name('upload');
 
@@ -46,6 +46,8 @@ Route::post('meals', 'MealControllerAPI@store');
 //Route::get('orders', 'OrderControllerAPI@index');
 Route::get('ordersWaiter', 'OrderControllerAPI@ordersWaiter');
 Route::post('orders', 'OrderControllerAPI@store');
+
+Route::get('mealsInfo', 'MealControllerAPI@dashInfo');
 
 //******************************************************************************
 //Route::middleware('auth:api')->
